requirements.txt
anyio==3.6.1
attrs==21.4.0
bcrypt==3.2.2
certifi==2022.6.15
cffi==1.15.1
# click==8.1.3
cryptography==37.0.4
distlib==0.3.5
fastapi==0.79.0
# filelock==3.7.1
greenlet==1.1.2
h11==0.13.0
idna==3.3
iniconfig==1.1.1
mysql-connector-python==8.0.29
packaging==21.3
# pipenv==2022.7.4
# platformdirs==2.5.2
pluggy==1.0.0
# protobuf==4.21.3
py==1.11.0
pycparser==2.21
pydantic==1.9.1
PyJWT==2.4.0
pyparsing==3.0.9
# pytest==7.1.2
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
# tomli==2.0.1
# typing_extensions==4.3.0
uvicorn==0.17.0
virtualenv==20.15.1
virtualenv-clone==0.5.7
scikit-learn==1.3.0
pandas==1.3.5